+++ CHANGELOG.md
@@ -8,7 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ##  [Unreleased]
 
-_tbd_
+
+### Added
+
+-   Allow to overwrite `language` parameter in method `API::request()`
+-   Enhance unit tests, mostly for testing fixed bugs in i-doit API add-on 1.9.1
 
 
 ##  [0.4] – 2018-02-21